Split out the regular expression models further
